pyproject.toml
[project]
name = "ecoacoustic-discovery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
dependencies = [
    "ipykernel>=6.30.0",
    "lap",
    "librosa>=0.11.0",
    "matplotlib>=3.10.3",
    "pandas>=2.3.1",
    "pyha-analyzer",
    "seaborn>=0.13.2",
    "torchcodec>=0.5",
    "tqdm>=4.67.1",
    "datasets<4.0.0",
    "bioacoustics-model-zoo>=0.12.0",
    "tf-keras>=2.19.0",
    "statsmodels>=0.14.5",
]

[tool.uv.sources]
pyha-analyzer = { git = "https://github.com/UCSD-E4E/pyha-analyzer-2.0.git" }
lap = { git = "https://github.com/gatagat/lap.git" }
